pyproject.toml
[project]
name = "spectask-mcp"
version = "0.1.11"
description = "Stdio MCP and CLI for Spectask-Jira workflows"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.2.0",
  "jira>=3.10.0",
  "requests[socks]>=2.32.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
spectask-mcp = "spectask_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spectask_mcp"]
